Initial commit
Initial commit for Uniborrow service Users.

Created a Java project using KumuluzEE framework. The project consits
of four modules: api, services, lib and modules.

The users service uses a postgres database to store the users.

A dockerfile is added - creation of docker image and running the
docker container was tested.